+++ Changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [v0.4.2-dev][Unreleased] - Unreleased
+## [v0.5.0] - 2024-10-24
 * Added support for Pluma and xed
 * Added support for moving tabs with Ctrl-Shift-Page Up and
   Ctrl-Shift-Page Down ([#13])
@@ -83,7 +83,7 @@
 * Initial release
 
 
-[Unreleased]: https://github.com/jefferyto/gedit-control-your-tabs/compare/v0.4.1...main
+[v0.5.0]: https://github.com/jefferyto/gedit-control-your-tabs/compare/v0.4.1...v0.5.0
 [v0.4.1]: https://github.com/jefferyto/gedit-control-your-tabs/compare/v0.4.0...v0.4.1
 [v0.4.0]: https://github.com/jefferyto/gedit-control-your-tabs/compare/v0.3.5...v0.4.0
 [v0.3.5]: https://github.com/jefferyto/gedit-control-your-tabs/compare/v0.3.4...v0.3.5